语音助手/requirements.txt
baidu-aip==2.2.13.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dlib==19.7.0
face-recognition==1.2.3
face-recognition-models==0.3.0
idna==2.8
numpy==1.16.2
opencv-python==4.0.0.21
Pillow==5.4.1
PyAudio==0.2.11
pygame==1.9.5
PyQt5==5.11.3
PyQt5-sip==4.19.15
pyqt5-tools==5.11.3.1.4
PyQtWebEngine==5.12.1
python-dotenv==0.10.1
pywin32==224
QScintilla==2.11.1
requests==2.21.0
scipy==1.2.1
SpeechRecognition==3.8.1
urllib3==1.24.1
wincertstore==0.2
